+++ AcademyChallengeApp/AcademyChallengeApp/Utilities/String+ToImage.swift
@@ -1,0 +1,22 @@
+//
+//  String+ToImage.swift
+//  AcademyChallengeApp
+//
+//  Created by Rúben Antunes on 07/10/2022.
+//
+
+import UIKit
+
+extension String {
+    func image() -> UIImage? {
+        let size = CGSize(width: 40, height: 40)
+        UIGraphicsBeginImageContextWithOptions(size, false, 0)
+        UIColor.white.set()
+        let rect = CGRect(origin: .zero, size: size)
+        UIRectFill(CGRect(origin: .zero, size: size))
+        (self as AnyObject).draw(in: rect, withAttributes: [.font: UIFont.systemFont(ofSize: 40)])
+        let image = UIGraphicsGetImageFromCurrentImageContext()
+        UIGraphicsEndImageContext()
+        return image
+    }
+}
